requirements.txt
requests
pyyaml
stomp.py
